timeout should be a double to support larger values

+++ android/src/main/java/com/ocetnik/timer/BackgroundTimerModule.java
@@ -75,7 +75,7 @@ public class BackgroundTimerModule extends ReactContextBaseJavaModule {
     }
 
     @ReactMethod
-    public void setTimeout(final int id, final int timeout) {
+    public void setTimeout(final int id, final double timeout) {
         Handler handler = new Handler();
         handler.postDelayed(new Runnable(){
             @Override
